拆分class
