+++ lib/Demo/hello-demo.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class Hello extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text(
+        "hello",
+        textDirection: TextDirection.ltr,
+        style: TextStyle(
+          fontSize: 40,
+          fontWeight: FontWeight.bold,
+          color: Colors.blue,
+        ),
+      ),
+    );
+  }
+}
