+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'model/post.dart';
+import './Demo/listview-demo.dart';
 
 void main() => runApp(
       App(), //自己定義的Widget
@@ -12,7 +12,9 @@ class App extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Home(),
-      theme: ThemeData(primarySwatch: Colors.yellow),
+      theme: ThemeData(
+        primarySwatch: Colors.yellow,
+      ),
     );
   }
 }
@@ -20,59 +22,17 @@ class App extends StatelessWidget {
 class Home extends StatelessWidget {
   const Home({Key? key}) : super(key: key);
 
-  Widget _listItemBuilder(BuildContext context, int index) {
-    return Container(
-      color: Colors.white,
-      margin: EdgeInsets.all(8.0),
-      child: Column(
-        children: <Widget>[
-          Image.network(posts[index].imageUrl),
-          SizedBox(
-            height: 16.0,
-          ),
-          Text(
-            posts[index].title,
-            style: Theme.of(context).textTheme.headline4,
-          ),
-          Text(
-            posts[index].author,
-            style: Theme.of(context).textTheme.headline6,
-          ),
-          SizedBox(height: 16.0,),
-        ],
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.grey[100],
       appBar: AppBar(
-        title: Text("塑蓋資料"),
+        title: Text("塑蓋"),
         elevation: 0.0,
       ),
-      body: ListView.builder(
-        itemCount: posts.length,
-        itemBuilder: _listItemBuilder,
-      ),
+      body: ListViewDemo(),
     );
   }
 }
 
-class Hello extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Text(
-        "hello",
-        textDirection: TextDirection.ltr,
-        style: TextStyle(
-          fontSize: 40,
-          fontWeight: FontWeight.bold,
-          color: Colors.blue,
-        ),
-      ),
-    );
-  }
-}
+
